Move HTTP utils to test-rpc
