+++ test-rpc/Cargo.toml
@@ -9,6 +9,7 @@ futures = "0.3"
 tokio = { version = "1.0", features = ["macros", "rt", "process", "time", "fs", "io-util", "rt-multi-thread"] }
 tarpc = { version = "0.30", features = ["tokio1", "serde-transport", "serde1"] }
 serde = "1.0"
+once_cell = "1.16.0"
 tokio-serde = { version = "0.8.0", features = ["json"] }
 serde_json = "1.0"
 bytes = "1.3.0"
@@ -16,6 +17,11 @@ err-derive = "0.3.1"
 log = "0.4.17"
 colored = "2.0.0"
 async-trait = "0.1.58"
+
+hyper = { version = "0.14.23", features = ["client", "http2"] }
+hyper-rustls = "0.23"
+tokio-rustls = "0.23"
+rustls-pemfile = "0.2"
 
 [dependencies.tokio-util]
 version = "0.7"
