+++ test-runner/Cargo.toml
@@ -18,10 +18,6 @@ serde = { version = "1.0" }
 serde_json = "1.0"
 tokio-serde = { version = "0.8.0", features = ["json"] }
 
-hyper = { version = "0.14.23", features = ["client", "http2"] }
-hyper-rustls = "0.23"
-tokio-rustls = "0.23"
-rustls-pemfile = "0.2"
 mullvad-management-interface = { git = "https://github.com/mullvad/mullvadvpn-app", branch = "main" }
 libc = "0.2"
